featuresericeApi - marzi and sarah updated

+++ src/main/java/com/mftplus/storage/controller/api/RoleApi.java
@@ -43,7 +43,7 @@ public class RoleApi {
     @DELETE
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response remove(@PathParam("id") Integer id) {
+    public Response remove(@PathParam("id") Long id) {
         try {
             return Response.ok().entity(roleService.remove(id)).build();
         } catch (Exception e) {
@@ -74,7 +74,7 @@ public class RoleApi {
         System.out.println(id);
 
         try {
-            return Response.ok().entity(roleService.findById(Integer.valueOf(id))).build();
+            return Response.ok().entity(roleService.findById(Long.valueOf(id))).build();
         } catch (Exception e) {
             System.out.println(e.getMessage());
             e.printStackTrace();
